Update bin files to allow easier setup/testing
Updated the setup command to include the require commands for initializing the
git submodules, as well as building the Argon2 C Library. Also added bin/test
to run the C test suite as well.

These were originally added to allow easy setup/testing in the Github Actions,
but also make local dev on the wrapper easier as well.